move run_playbook helpers to conftest
makes them usable outside of test_crud
